CRUD operation for User. and use JWT.

+++ applications/views.py
@@ -7,6 +7,8 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.parsers import JSONParser
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticated
+from rest_framework_simplejwt.authentication import JWTAuthentication
 
 # def applicationsList(request):
 #     app_ids = Application.objects.all()
@@ -15,6 +17,9 @@ from rest_framework import status
 
 
 class ApplicationListView(APIView):
+    permission_classes = [IsAuthenticated,]
+    authentication_classes = [JWTAuthentication,]
+
     '''Get List of applications and post also'''
     def get(self, request):
         '''Read application using DRF'''
@@ -24,6 +29,7 @@ class ApplicationListView(APIView):
 
     def post(self, request):
         '''Create application using DRF'''
+        # import pdb;pdb.set_trace()
         app_serializer = ApplicationSerializer(data=request.data)
         if app_serializer.is_valid():
             app_serializer.save()
@@ -49,14 +55,18 @@ class ApplicationDetailsView(APIView):
     def delete(self, request, pk):
         '''Delete single object of Application'''
         application = self.get_application_obj(pk)
-        application.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user.id == application.user.id:
+            application.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
+
 
     def put(self, request, pk):
         '''Update single object of Application'''
         application = self.get_application_obj(pk)
-        app_serializer = ApplicationSerializer(application, data=request.data)
-        if app_serializer.is_valid():
-            app_serializer.save()
-            return Response(app_serializer.data)
+        if request.user.id == application.user.id:
+            app_serializer = ApplicationSerializer(application, data=request.data)
+            if app_serializer.is_valid():
+                app_serializer.save()
+                return Response(app_serializer.data)
         return Response(app_serializer.errors ,status.HTTP_400_BAD_REQUEST)
